requirements.txt
discord.py
torch
torchvision
requests
Pillow
flask==2.2.3
gunicorn==20.1.0
